When an output selection cancels an entire particle type, set NumParticleThisFile to 0 (for that type) in the snapshot header

+++ src/fof.h
@@ -87,7 +87,7 @@ struct fof_props {
   /* ------------  Group properties ----------------- */
 
   /*! Number of groups */
-  int num_groups;
+  long long num_groups;
 
   /*! Number of local black holes that belong to groups whose roots are on a
    * different node. */
